requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
python-telegram-bot==13.14
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tornado==6.1
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
